requirements.txt
discord.py>=2.3.2
google-genai>=0.3.0
python-dotenv>=1.0.0
tenacity>=8.2.3